persist replay update state in ChannelActorData
