+++ crates/fiber-lib/src/fiber/tests/settle_tlc_set_command_tests.rs
@@ -298,16 +298,16 @@ fn create_test_channel_state_with_tlc(
             latest_commitment_transaction: None,
             reestablishing: false,
             last_revoke_ack_msg: None,
+            pending_replay_updates: vec![],
+            last_was_revoke: false,
             created_at: SystemTime::now(),
         },
-        pending_replay_updates: vec![],
         waiting_peer_response: None,
         network: None,
         scheduled_channel_update_handle: None,
         pending_notify_settle_tlcs: vec![],
         defer_peer_tlc_updates: false,
         deferred_peer_tlc_updates: std::collections::VecDeque::new(),
-        last_was_revoke: false,
         ephemeral_config: Default::default(),
         private_key: None,
     }
