+++ crates/fiber-lib/src/store/sample/sample_channel.rs
@@ -122,17 +122,16 @@ impl ChannelActorState {
                 shutdown_transaction_hash: None,
                 reestablishing: false,
                 last_revoke_ack_msg: None,
+                pending_replay_updates: vec![],
+                last_was_revoke: false,
                 created_at: deterministic_time(0),
             },
-            // #[serde(skip)] fields — not serialized, use defaults
-            pending_replay_updates: vec![],
             waiting_peer_response: None,
             network: None,
             scheduled_channel_update_handle: None,
             pending_notify_settle_tlcs: vec![],
             defer_peer_tlc_updates: false,
             deferred_peer_tlc_updates: VecDeque::new(),
-            last_was_revoke: false,
             ephemeral_config: Default::default(),
             private_key: None,
         }
@@ -359,17 +358,16 @@ impl ChannelActorState {
                 shutdown_transaction_hash: Some(ckb_types::H256(deterministic_hash(seed, 950))),
                 reestablishing: true,
                 last_revoke_ack_msg: Some(revoke_and_ack),
+                pending_replay_updates: vec![],
+                last_was_revoke: true,
                 created_at: deterministic_time(3600),
             },
-            // #[serde(skip)] fields — not serialized, use defaults
-            pending_replay_updates: vec![],
             waiting_peer_response: None,
             network: None,
             scheduled_channel_update_handle: None,
             pending_notify_settle_tlcs: vec![],
             defer_peer_tlc_updates: false,
             deferred_peer_tlc_updates: VecDeque::new(),
-            last_was_revoke: true,
             ephemeral_config: Default::default(),
             private_key: None,
         }
